move ReportStatus class
