+++ backend/src/main/java/pl/edu/pw/pap/review/ReviewService.java
@@ -6,7 +6,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import pl.edu.pw.pap.comment.ForbiddenException;
-import pl.edu.pw.pap.comment.report.ReportStatus;
+import pl.edu.pw.pap.report.ReportStatus;
 import pl.edu.pw.pap.course.Course;
 import pl.edu.pw.pap.course.CourseRepository;
 import pl.edu.pw.pap.course.CourseNotFoundException;
