+++ backend/src/main/java/pl/edu/pw/pap/review/report/ReviewReport.java
@@ -3,7 +3,7 @@ package pl.edu.pw.pap.review.report;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-import pl.edu.pw.pap.comment.report.ReportStatus;
+import pl.edu.pw.pap.report.ReportStatus;
 import pl.edu.pw.pap.review.Review;
 import pl.edu.pw.pap.user.User;
